Remove stream support for now since it doesn't get along with nginx.
nginx, as built by default doesn't support
proxying PUT or POST requests without a
Content-Length header set. The chunkin module adds
support for such (and has been verified in my
tests), but isn't packaged for Ubuntu by default.
For now we'll take the simple route and remove
support in the bindings unless there are requests
for it in the future.