Update to latest Jeweler, regenerate gem spec
The current version of Jeweler now uses Gemfiles a la Bundler.  One tricky
thing to note for the future is that 'rake gemspec' totally fails if the
gem's root directory isn't the same as the git root directory, so we have to
copy the dir to another place (e.g. /tmp) to create the specfile.